requirements.txt
sentence-transformers
flask
langchain
pypdf 
python-dotenv
pinecone-client
langchain-community
langchain-pinecone
langchain-experimental
groq
chainlit
-e .